+++ test_push.py
@@ -13,9 +13,9 @@ print(t.call(python_version))
 print(t.call(check_output, ['ls', '-l', res['remote_path']]))
 
 g = Group([t, Docker('test-2')])
-for host, res in g.put('/usr/share/common-licenses/GPL').successful():
+for host, res in g.put('/usr/share/common-licenses/GPL').iteritems():
     print(host, res)
 
-for host, res in g.call(check_output, ['ls', '-l', '/tmp']).successful():
+for host, res in g.call(check_output, ['ls', '-l', '/tmp']).iteritems():
     print(host)
     print(res)